requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy>=1.4.36,<2.0.0
alembic==1.13.1

# Vector DB
qdrant-client==1.7.1

# NLP / Embeddings
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# dbt
dbt-core==1.7.4
dbt-postgres==1.7.4

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# Object storage
boto3==1.34.14
minio==7.2.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.2.0

# Code quality
ruff==0.1.14
black==23.12.1

